requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-elasticsearch==0.3.0
fastapi>=0.110
uvicorn[standard]>=0.23
python-multipart>=0.0.6
pydantic-settings>=2.0
sentence-transformers>=2.2
torch>=2.0
easyocr>=1.7
opencv-python>=4.8
Pillow>=9.5
numpy>=1.23
langdetect>=1.0.9
watchdog>=4.0
pymupdf>=1.23
pdfminer.six>=20221105
unstructured>=0.12
elastic-transport>=8.4
elasticsearch>=8.11
typer>=0.9
python-dotenv>=1.0
rapidfuzz>=3.6
scikit-learn>=1.3
scipy>=1.10
# Production dependencies
gunicorn>=21.0
redis>=5.0
slowapi>=0.1.9
cachetools>=5.3
# NLP for accuracy
spacy>=3.7
indic-nlp-library>=0.92
# LLM Provider (OpenAI only)
langchain-openai>=0.1.0

